Fixed compilation warnings

+++ DbAdapterTestUtilities/Stubs/StubPrimaryKey.cpp
@@ -27,7 +27,7 @@ namespace systelab { namespace db { namespace test_utility {
 
 	unsigned int StubPrimaryKey::getFieldsCount() const
 	{
-		return m_fields.size();
+		return (unsigned int) m_fields.size();
 	}
 
 	const IField& StubPrimaryKey::getField(unsigned int index) const
@@ -44,7 +44,7 @@ namespace systelab { namespace db { namespace test_utility {
 
 	const IField& StubPrimaryKey::getField(const std::string& fieldName) const
 	{
-		unsigned int nFields = m_fields.size();
+		unsigned int nFields = (unsigned int) m_fields.size();
 		for (unsigned int i = 0; i < nFields; i++)
 		{
 			if (m_fields[i]->getName() == fieldName)
