+++ DbAdapterTestUtilities/Stubs/StubRecord.cpp
@@ -15,7 +15,7 @@ namespace systelab { namespace db { namespace test_utility {
 	StubRecord::StubRecord(const StubRecord& other)
 		:m_fieldValues()
 	{
-		unsigned int nFieldValues = other.m_fieldValues.size();
+		unsigned int nFieldValues = (unsigned int) other.m_fieldValues.size();
 		for (unsigned int i = 0; i < nFieldValues; i++)
 		{
 			m_fieldValues.push_back(std::unique_ptr<StubFieldValue>(new StubFieldValue(*other.m_fieldValues[i].get())));
@@ -26,7 +26,7 @@ namespace systelab { namespace db { namespace test_utility {
 
 	StubRecord::StubRecord(std::vector< std::unique_ptr<StubFieldValue> >& fieldValues)
 	{
-		unsigned int nFieldValues = fieldValues.size();
+		unsigned int nFieldValues = (unsigned int) fieldValues.size();
 		for (unsigned int i = 0; i < nFieldValues; i++)
 		{
 			m_fieldValues.push_back(std::move(fieldValues[i]));
@@ -39,7 +39,7 @@ namespace systelab { namespace db { namespace test_utility {
 
 	unsigned int StubRecord::getFieldValuesCountStub() const
 	{
-		return m_fieldValues.size();
+		return (unsigned int) m_fieldValues.size();
 	}
 
 	db::IFieldValue& StubRecord::getFieldValueByIndexStub(unsigned int index) const
@@ -49,7 +49,7 @@ namespace systelab { namespace db { namespace test_utility {
 
 	db::IFieldValue& StubRecord::getFieldValueByNameStub(const std::string& name) const
 	{
-		unsigned int nFieldValues = m_fieldValues.size();
+		unsigned int nFieldValues = (unsigned int) m_fieldValues.size();
 		for (unsigned int i = 0; i < nFieldValues; i++)
 		{
 			std::string fieldName = m_fieldValues[i]->getField().getName();
@@ -64,7 +64,7 @@ namespace systelab { namespace db { namespace test_utility {
 
 	bool StubRecord::hasFieldValueStub(const std::string& name) const
 	{
-		unsigned int nFieldValues = m_fieldValues.size();
+		unsigned int nFieldValues = (unsigned int) m_fieldValues.size();
 		for (unsigned int i = 0; i < nFieldValues; i++)
 		{
 			std::string fieldName = m_fieldValues[i]->getField().getName();
@@ -79,7 +79,7 @@ namespace systelab { namespace db { namespace test_utility {
 
 	StubRecord& StubRecord::operator= (const StubRecord& other)
 	{
-		unsigned int nFieldValues = other.m_fieldValues.size();
+		unsigned int nFieldValues = (unsigned int) other.m_fieldValues.size();
 		for (unsigned int i = 0; i < nFieldValues; i++)
 		{
 			m_fieldValues.push_back(std::unique_ptr<StubFieldValue>(new StubFieldValue(*other.m_fieldValues[i].get())));
